// infra/lib/types/stack-options.ts
export type StackOptions = {
  defaultRegion: string;
  stackNamePrefix: string;
  stackName: string;
  codeCommitAccount: string;
  toolsAccount: string;
  reposName: string;
  devAccount: string;
  qaAccount: string;
  stagingAccount: string;
  productionAccount: string;
  cdkBootstrapQualifier: string;
  pipelineName: string;
  apiStageName: string;
  apiName: string;
};